# api/swagger.yaml
definitions:
...
  v4.Budget:
    properties:
      createdAt:
        description: Time the resource was created
        example: "2022-04-02T19:28:44.491514Z"
        type: string
      currency:
        description: The currency for the budget
        example: €
        type: string
      deletedAt:
        description: Time the resource was marked as deleted
        example: "2022-04-22T21:01:05.058161Z"
        type: string
      id:
        description: UUID for the resource
        example: 65392deb-5e92-4268-b114-297faad6cdce
        type: string
      links:
        $ref: '#/definitions/v4.BudgetLinks'
      name:
        description: Name of the budget
        example: Morre's Budget
        type: string
      note:
        description: A longer description of the budget
        example: My personal expenses
        type: string
      updatedAt:
        description: Last time the resource was updated
        example: "2022-04-17T20:14:01.048145Z"
        type: string
    type: object
...
  v4.BudgetCreateResponse:
    properties:
      data:
        description: List of created Budgets
        items:
          $ref: '#/definitions/v4.BudgetResponse'
        type: array
      error:
        description: 'FIMXE: make this *error for ALL responses'
        example: the specified resource ID is not a valid UUID
        type: string
    type: object
  v4.BudgetEditable:
    properties:
      currency:
        description: The currency for the budget
        example: €
        type: string
      name:
        description: Name of the budget
        example: Morre's Budget
        type: string
      note:
        description: A longer description of the budget
        example: My personal expenses
        type: string
    type: object
  v4.BudgetLinks:
    properties:
      accounts:
        description: Accounts for this budget
        example: https://example.com/api/v4/accounts?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf
        type: string
      categories:
        description: Categories for this budget
        example: https://example.com/api/v4/categories?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf
        type: string
      envelopes:
        description: Envelopes for this budget
        example: https://example.com/api/v4/envelopes?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf
        type: string
      month:
        description: This uses 'YYYY-MM' for clients to replace with the actual year
          and month.
        example: https://example.com/api/v4/months?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf&month=YYYY-MM
        type: string
      self:
        description: The budget itself
        example: https://example.com/api/v4/budgets/550dc009-cea6-4c12-b2a5-03446eb7b7cf
        type: string
      transactions:
        description: Transactions for this budget
        example: https://example.com/api/v4/transactions?budget=550dc009-cea6-4c12-b2a5-03446eb7b7cf
        type: string
    type: object
  v4.BudgetListResponse:
    properties:
      data:
        description: List of budgets
        items:
          $ref: '#/definitions/v4.Budget'
        type: array
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/v4.Pagination'
        description: Pagination information
    type: object
  v4.BudgetResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/v4.Budget'
        description: Data for the budget
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
    type: object
...
  v4.Pagination:
    properties:
      count:
        description: The amount of records returned in this response
        example: 25
        type: integer
      limit:
        description: The maximum amount of resources to return for this request
        example: 25
        type: integer
      offset:
        description: The offset for the first record returned
        example: 50
        type: integer
      total:
        description: The total number of resources matching the query
        example: 827
        type: integer
    type: object
...
  v4.httpError:
    properties:
      error:
        example: An ID specified in the query string was not a valid UUID
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v4/budgets:
    get:
      description: Returns a list of budgets
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by note
        in: query
        name: note
        type: string
      - description: Filter by currency
        in: query
        name: currency
        type: string
      - description: Search for this text in name and note
        in: query
        name: search
        type: string
      - description: The offset of the first Budget returned. Defaults to 0.
        in: query
        name: offset
        type: integer
      - description: Maximum number of Budgets to return. Defaults to 50.
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.BudgetListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.BudgetListResponse'
      summary: List budgets
      tags:
      - Budgets
    options:
      description: Returns an empty response with the HTTP Header "allow" set to the
        allowed HTTP verbs
      responses:
        "204":
          description: No Content
      summary: Allowed HTTP verbs
      tags:
      - Budgets
    post:
      consumes:
      - application/json
      description: Creates a new budget
      parameters:
      - description: Budget
        in: body
        name: budget
        required: true
        schema:
          items:
            $ref: '#/definitions/v4.BudgetEditable'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v4.BudgetCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.BudgetCreateResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.BudgetCreateResponse'
      summary: Create budget
      tags:
      - Budgets
  /v4/budgets/{id}:
    delete:
      description: Deletes a budget
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.httpError'
      summary: Delete budget
      tags:
      - Budgets
    get:
      description: Returns a specific budget
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.BudgetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.BudgetResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.BudgetResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.BudgetResponse'
      summary: Get budget
      tags:
      - Budgets
    options:
      description: Returns an empty response with the HTTP Header "allow" set to the
        allowed HTTP verbs
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.httpError'
      summary: Allowed HTTP verbs
      tags:
      - Budgets
    patch:
      consumes:
      - application/json
      description: Update an existing budget. Only values to be updated need to be
        specified.
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      - description: Budget
        in: body
        name: budget
        required: true
        schema:
          $ref: '#/definitions/v4.BudgetEditable'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.BudgetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.BudgetResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.BudgetResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.BudgetResponse'
      summary: Update budget
      tags:
      - Budgets
...
swagger: "2.0"